refactor so less tests are used
