+++ .gitignore
@@ -22,3 +22,4 @@ modules/utilities.js
 modules/mod_sources/github_source.js
 modules/mod_sources/jsonlist_source.js
 modules/mod_sources/mod_source_base.js
+modules/mod_sources/github_collection_source.js
